backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
